+++ shoop/admin/modules/contact_groups/views/edit.py
@@ -8,9 +8,11 @@
 
 from __future__ import unicode_literals
 
+from django.core.urlresolvers import reverse_lazy
 from django.db.transaction import atomic
 
 from shoop.admin.form_part import FormPartsViewMixin, SaveFormPartsMixin
+from shoop.admin.toolbar import get_default_edit_toolbar
 from shoop.admin.utils.views import CreateOrUpdateView
 from shoop.core.models import ContactGroup
 
@@ -26,3 +28,9 @@ class ContactGroupEditView(SaveFormPartsMixin, FormPartsViewMixin, CreateOrUpdat
     @atomic
     def form_valid(self, form):
         return self.save_form_parts(form)
+
+    def get_toolbar(self):
+        save_form_id = self.get_save_form_id()
+        object = self.get_object()
+        delete_url = reverse_lazy("shoop_admin:contact-group.delete", kwargs={"pk": object.pk})
+        return get_default_edit_toolbar(self, save_form_id, delete_url=delete_url if object.can_delete() else None)